fix: to float; parser tests pass

+++ src/Execution/AstEvaluator.cs
@@ -47,7 +47,7 @@ public class AstEvaluator : IAstVisitor
         RuntimeValue value = expression.Value switch
         {
             int i => new RuntimeValue(i),
-            double d => new RuntimeValue(d),
+            float d => new RuntimeValue(d),
             string s => new RuntimeValue(s),
             bool b => new RuntimeValue(b),
             _ => throw new Exception($"Unknown expression type met while trying to visit {nameof(LiteralExpression)}"),
@@ -156,9 +156,9 @@ public class AstEvaluator : IAstVisitor
                 RuntimeValueType valueType = value.GetValueType();
                 RuntimeValueType variableType = ToRuntimeValueType(declaration.VariableType);
 
-                if (valueType == RuntimeValueType.Int && variableType == RuntimeValueType.Double)
+                if (valueType == RuntimeValueType.Int && variableType == RuntimeValueType.Float)
                 {
-                    value = new RuntimeValue(value.ToDouble());
+                    value = new RuntimeValue(value.ToFloat());
                 }
                 else if (valueType != variableType)
                 {
@@ -343,9 +343,9 @@ public class AstEvaluator : IAstVisitor
                 RuntimeValueType valueType = value.GetValueType();
                 RuntimeValueType returnType = ToRuntimeValueType(function.Type);
 
-                if (valueType == RuntimeValueType.Int && returnType == RuntimeValueType.Double)
+                if (valueType == RuntimeValueType.Int && returnType == RuntimeValueType.Float)
                 {
-                    values.Push(new RuntimeValue(values.Pop().ToDouble()));
+                    values.Push(new RuntimeValue(values.Pop().ToFloat()));
                 }
                 else if (valueType != returnType)
                 {
@@ -469,9 +469,9 @@ public class AstEvaluator : IAstVisitor
             RuntimeValue value = environment.ReadValue(variableType);
             RuntimeValueType valueType = value.GetValueType();
 
-            if (valueType == RuntimeValueType.Int && variableType == RuntimeValueType.Double)
+            if (valueType == RuntimeValueType.Int && variableType == RuntimeValueType.Float)
             {
-                context.AssignVariable(variableName, new RuntimeValue(value.ToDouble()));
+                context.AssignVariable(variableName, new RuntimeValue(value.ToFloat()));
             }
             else if (valueType != variableType)
             {
@@ -585,7 +585,7 @@ public class AstEvaluator : IAstVisitor
             case UnaryOperation.Minus:
             case UnaryOperation.Increment:
             case UnaryOperation.Decrement:
-                return value is RuntimeValueType.Int or RuntimeValueType.Double;
+                return value is RuntimeValueType.Int or RuntimeValueType.Float;
             case null:
                 return true;
             default:
@@ -622,7 +622,7 @@ public class AstEvaluator : IAstVisitor
 
     private bool IsNumericType(RuntimeValueType value)
     {
-        return value is RuntimeValueType.Int or RuntimeValueType.Double;
+        return value is RuntimeValueType.Int or RuntimeValueType.Float;
     }
 
     private RuntimeValueType ToRuntimeValueType(VariableType value)
@@ -630,7 +630,7 @@ public class AstEvaluator : IAstVisitor
         return value switch
         {
             VariableType.Int => RuntimeValueType.Int,
-            VariableType.Double => RuntimeValueType.Double,
+            VariableType.Double => RuntimeValueType.Float,
             VariableType.String => RuntimeValueType.String,
             VariableType.Boolean => RuntimeValueType.Boolean,
             _ => throw new Exception($"Unexpected value type: {value}"),
